NH-3817 Test Case and Proposed Fix
Includes a test that causes Merge to fail on a graph containing
transient entities with multiple cascade paths.
NOTE: The test may not consistently fail in 100% of cases. Please see
NH-3817 for explanation.
Fixes DefaultMergeEventListener to ensure that all transient entities in a
graph with multiple cascade paths get merged regardless of the order of
entities in transient copy cache.